refactor: move LongPollingIT and ProcessInstanceMigrationClusteredTest to engine/client/multipartition package
These tests are about multi-partition client behaviour (long polling across
partitions, process instance migration between partitions), not about cluster
infrastructure. Moving them to io.camunda.zeebe.it.engine.client.multipartition
attributes them to the correct team.
