+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/engine/client/multipartition/LongPollingIT.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.it.cluster.clustering;
+package io.camunda.zeebe.it.engine.client.multipartition;
 
 import static org.assertj.core.api.Assertions.assertThat;
